# frontend/src/api/openapi.yaml
openapi: 3.0.3
info:
  title: InstaGrader API
  version: 1.0.0
  description: API for the InstaGrader essay grading platform
paths:
...
  /api/auth/change-password/:
    post:
      operationId: auth_change_password
      description: Change the authenticated user's password.
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePassword'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ChangePassword'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ChangePassword'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePasswordSuccess'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
              examples:
                IncorrectOldPassword:
                  value:
                    old_password:
                    - Old password is incorrect.
                  summary: Incorrect old password
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
          description: ''
...
  /api/auth/logout/:
    post:
      operationId: auth_logout
      description: Logout by clearing authentication cookies.
      tags:
      - auth
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
          description: ''
...
components:
  schemas:
...
    ChangePassword:
      type: object
      description: Serializer for password change.
      properties:
        old_password:
          type: string
        new_password:
          type: string
      required:
      - new_password
      - old_password
    ChangePasswordSuccess:
      type: object
      properties:
        detail:
          type: string
          default: Password changed successfully.
...
    LogoutSuccess:
      type: object
      properties:
        detail:
          type: string
          default: Successfully logged out.
...
    UnauthorizedResponse:
      type: object
      properties:
        detail:
          type: string
          default: Authentication credentials were not provided.
        code:
          type: string
          default: not_authenticated
...
    ValidationErrorResponse:
      type: object
      properties:
        field_name:
          type: array
          items:
            type: string
          description: Each field maps to a list of validation error strings.
      required:
      - field_name
...